indentation on Composer and others stuff
Added automatic indenting on Composer

FormBuilder was updated to know about that automatic indenting and it
skips over spaces that won't be needed (&nbsp; kind of stuff)

FormBuilder::sessionValue is now a static function

FormBuilder::__construct has a new signature to take the attributes you
want in the FORM element. This is now the way to include
'enctype'='multipart/form-data', or 'id'=XYZ or whatever.

Fixed a bug with getTimeDisplay that was chopping off too much of the
time.

The content of a button is now specified with 'content' instead of
'display' in the $butts array parameter