Updated `composeActionLinks` in /wComposer to take a new KEY_HREF with values that will be passed to `w_compose_url`. It also has new KEY_WRAP_CLASS that can be used to set the class on the <P> tag that wraps the action link <A> tag.
/wForm takes an optional $cp parameter to set the form builder's
internal composer

Added $path parameter to `w_bailout` to explicitly set the path if
desired (if different from the current script)

Changed the `wrap` method to test first if the $test parameter is an
array. Before it was testing for boolean and then assuming every other
case was an array.

Updated the `dcode` method to utilize all ASCII characters between 32
(space) and 126 (tilde) instead of just the letters and digits. This
means our shifting can be done with one regular expression instead of
three.

In /wTemplate, the internal composer is constructed from the class name
stored at the constant CLASS_CP, allowing sub-classes to override the
const and provide a different one.